refactor: Consolidate hash table builders for semi-join and anti-join
Extract duplicate hash table building functions from hash_semi_join.rs and
hash_anti_join.rs into a shared implementation in hash_join/build.rs.

Changes:
- Add build_existence_hash_table_sequential() and build_existence_hash_table_parallel()
  to hash_join/build.rs for building HashMap<SqlValue, ()> existence tables
- Update hash_semi_join.rs to use shared builders, removing ~83 lines of duplicate code
- Update hash_anti_join.rs to use shared builders, removing ~83 lines of duplicate code
- Use existing combine_rows() helper from hash_join/mod.rs instead of local helpers

This reduces code duplication by ~130 lines while maintaining identical behavior.
The existence hash table builders only track key presence (not row indices), which
is more memory-efficient for semi-join and anti-join operations.

Closes #2830
