+++ crates/vibesql-executor/src/select/join/hash_join/build.rs
@@ -9,6 +9,9 @@ use crate::select::parallel::ParallelConfig;
 #[cfg(feature = "simd")]
 use crate::simd::hashing::simd_hash_sqlvalue_batch;
 
+use crate::errors::ExecutorError;
+use crate::timeout::{TimeoutContext, CHECK_INTERVAL};
+
 /// Build hash table sequentially using indices (fallback for small inputs)
 ///
 /// Returns a map from join key to row indices, avoiding storing row references
@@ -195,4 +198,100 @@ pub(super) fn build_hash_table_parallel_simd(
             }
             acc
         })
+}
+
+// ============================================================================
+// Existence Hash Table Builders (for semi-join and anti-join)
+// ============================================================================
+//
+// These functions build hash tables that only track key existence (HashMap<SqlValue, ()>)
+// rather than storing row indices. This is more memory-efficient for semi-join and
+// anti-join operations where we only need to know if a key exists, not which rows match.
+
+/// Build existence hash table sequentially (stores only keys, not indices)
+///
+/// For semi-join and anti-join, we only need to know if a key exists, not track all
+/// matching rows. This saves memory compared to inner join's Vec<usize> storage.
+pub(crate) fn build_existence_hash_table_sequential(
+    build_rows: &[vibesql_storage::Row],
+    build_col_idx: usize,
+    timeout_ctx: &TimeoutContext,
+) -> Result<HashMap<vibesql_types::SqlValue, ()>, ExecutorError> {
+    let mut hash_table: HashMap<vibesql_types::SqlValue, ()> = HashMap::new();
+    for (idx, row) in build_rows.iter().enumerate() {
+        // Check timeout periodically during build phase
+        if idx % CHECK_INTERVAL == 0 {
+            timeout_ctx.check()?;
+        }
+        let key = row.values[build_col_idx].clone();
+        // Skip NULL values - they never match in equi-joins
+        if key != vibesql_types::SqlValue::Null {
+            hash_table.insert(key, ());
+        }
+    }
+    Ok(hash_table)
+}
+
+/// Build existence hash table in parallel (for semi-join/anti-join)
+///
+/// Algorithm (when parallel feature enabled):
+/// 1. Divide build_rows into chunks (one per thread)
+/// 2. Each thread builds a local hash table from its chunk (no synchronization)
+/// 3. Merge partial hash tables sequentially (fast because we only store keys)
+///
+/// Performance: 3-6x speedup on large joins (50k+ rows) with 4+ cores
+/// Note: Falls back to sequential when parallel feature is disabled
+pub(crate) fn build_existence_hash_table_parallel(
+    build_rows: &[vibesql_storage::Row],
+    build_col_idx: usize,
+    timeout_ctx: &TimeoutContext,
+) -> Result<HashMap<vibesql_types::SqlValue, ()>, ExecutorError> {
+    #[cfg(feature = "parallel")]
+    {
+        let config = ParallelConfig::global();
+
+        // Use sequential fallback for small inputs
+        if !config.should_parallelize_join(build_rows.len()) {
+            return build_existence_hash_table_sequential(build_rows, build_col_idx, timeout_ctx);
+        }
+
+        // Check timeout before parallel execution (can't check mid-parallel easily)
+        timeout_ctx.check()?;
+
+        // Phase 1: Parallel build of partial hash tables
+        // Each thread processes a chunk and builds its own hash table
+        let chunk_size = (build_rows.len() / config.num_threads).max(1000);
+        let partial_tables: Vec<HashMap<_, ()>> = build_rows
+            .par_chunks(chunk_size)
+            .map(|chunk| {
+                let mut local_table: HashMap<vibesql_types::SqlValue, ()> = HashMap::new();
+                for row in chunk.iter() {
+                    let key = row.values[build_col_idx].clone();
+                    if key != vibesql_types::SqlValue::Null {
+                        local_table.insert(key, ());
+                    }
+                }
+                local_table
+            })
+            .collect();
+
+        // Check timeout after parallel build
+        timeout_ctx.check()?;
+
+        // Phase 2: Sequential merge of partial tables
+        // This is fast because we only need to insert keys, not append vectors
+        let result = partial_tables.into_iter().fold(HashMap::new(), |mut acc, partial| {
+            for (key, _) in partial {
+                acc.insert(key, ());
+            }
+            acc
+        });
+        Ok(result)
+    }
+
+    #[cfg(not(feature = "parallel"))]
+    {
+        // Always use sequential build when parallel feature is disabled
+        build_existence_hash_table_sequential(build_rows, build_col_idx, timeout_ctx)
+    }
 }
