+++ crates/vibesql-executor/src/select/join/hash_join/mod.rs
@@ -20,6 +20,9 @@ mod tests;
 pub(super) use inner::hash_join_inner;
 pub(super) use outer::hash_join_left_outer;
 
+// Re-export existence hash table builders for semi-join and anti-join
+pub(super) use build::build_existence_hash_table_parallel;
+
 // Re-export FromResult type for use in submodules
 pub(super) use super::FromResult;
